Add headers to cli
