+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cdevent"
-version = "0.0.1"
+version = "0.2.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
